refactor: fix the package structure to fit the MVC Pattern
