+++ src/main/kotlin/racingcar/domain/RandomGenerator.kt
@@ -1,5 +1,6 @@
-package racingcar.util
+package racingcar.domain
 
+import racingcar.util.RacingRandomSetting
 import kotlin.random.Random
 
 class RandomGenerator(private val setting: RandomSetting = RacingRandomSetting()) {
