+++ src/main/kotlin/racingcar/domain/model/GameSetting.kt
@@ -1,3 +1,3 @@
-package racingcar.model
+package racingcar.domain.model
 
 data class GameSetting(val names: List<String>, val turnCount: Int)
