+++ src/main/kotlin/racingcar/domain/model/InputResult.kt
@@ -1,3 +1,3 @@
-package racingcar.model
+package racingcar.domain.model
 
 data class InputResult(val names: String?, val turnCount: String?)
