+++ src/main/kotlin/racingcar/util/RacingRandomSetting.kt
@@ -1,5 +1,7 @@
 package racingcar.util
 
+import racingcar.domain.RandomSetting
+
 class RacingRandomSetting : RandomSetting {
     override val from: Int
         get() = 0
