+++ src/test/kotlin/racingcar/domain/RandomGeneratorTest.kt
@@ -1,8 +1,7 @@
-package racingcar
+package racingcar.domain
 
 import org.assertj.core.api.Assertions
 import org.junit.jupiter.api.Test
-import racingcar.util.RandomGenerator
 
 class RandomGeneratorTest {
     @Test
