+++ src/test/kotlin/racingcar/domain/TestDouble.kt
@@ -1,7 +1,7 @@
-package racingcar
+package racingcar.domain
 
-import racingcar.model.Car
-import racingcar.rule.RacingRule
+import racingcar.domain.model.Car
+import racingcar.domain.rule.RacingRule
 
 class AlwaysMoveRule : RacingRule {
     override fun canMove(): Boolean = true
