+++ src/test/kotlin/racingcar/util/InputParserTest.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.util
 
 import org.assertj.core.api.Assertions
 import org.junit.jupiter.api.Test
